server/requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
supabase==2.7.4
python-dotenv==1.0.1
